project_tests/MPEI-master/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-ckeditor==6.5.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-modeltranslation==0.18.7
Pillow==9.3.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.6
